% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_alt_summ}
\alias{extract_alt_summ}
\title{Extract ALT summary as a data frame}
\usage{
extract_alt_summ(alt_obj, digits = 4)
}
\arguments{
\item{alt_obj}{Object returned by \code{WeibullR.ALT::alt.fit()}.}

\item{digits}{Number of decimal places to round numeric values.}
}
\value{
A two-column data frame with \code{Param} and \code{Value}.
}
\description{
Extract ALT summary as a data frame
}
