% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliaShiny.R
\name{ReliaShiny}
\alias{ReliaShiny}
\title{A Shiny Reliability Analysis App.}
\usage{
ReliaShiny()
}
\value{
This function does not return a value.
}
\description{
This function launches a Shiny application for reliability analysis.
}
\examples{
if (interactive()) {
  ReliaShiny()
}
}
\seealso{
\url{https://paulgovan.github.io/ReliaShiny/}
}
